Move repl logic to core
