+++ libs/core/src/assert.ts
@@ -1,0 +1,3 @@
+export function assertUnreachable(_v: never): never {
+  throw new Error("Unreachable code reached. Uh oh.");
+}
