+++ libs/core/src/index.ts
@@ -1,3 +1,2 @@
-export function assertUnreachable(_v: never): never {
-  throw new Error("Unreachable code reached. Uh oh.");
-}
+export * from "./assert.js";
+export * from "./repl.js";
